fix: remove absolute Claude hook commands
