+++ src/hooks/mod.rs
@@ -11,3 +11,35 @@ pub mod permissions;
 pub mod rewrite_cmd;
 pub mod trust;
 pub mod verify_cmd;
+
+pub fn is_claude_hook_command(command: &str) -> bool {
+    let parts = crate::discover::lexer::shell_split(command);
+    let [binary, hook, claude] = parts.as_slice() else {
+        return false;
+    };
+
+    let binary_name = binary.rsplit(['/', '\\']).next().unwrap_or(binary);
+
+    binary_name == "rtk" && hook == "hook" && claude == "claude"
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn claude_hook_command_matches_bare_and_absolute_rtk() {
+        assert!(is_claude_hook_command("rtk hook claude"));
+        assert!(is_claude_hook_command("/opt/homebrew/bin/rtk hook claude"));
+        assert!(is_claude_hook_command(
+            "\"/opt/homebrew/bin/rtk\" hook claude"
+        ));
+    }
+
+    #[test]
+    fn claude_hook_command_rejects_other_commands() {
+        assert!(!is_claude_hook_command("not-rtk hook claude"));
+        assert!(!is_claude_hook_command("/opt/homebrew/bin/rtk hook cursor"));
+        assert!(!is_claude_hook_command("echo rtk hook claude"));
+    }
+}
